refactor(models): 폴더명 수정
