+++ src/models/Lotto.js
